or1200-generic: verilator/tb: refactor parseArgs()
This reverts the change from if-else to if-continue introduced in
44ee3d8836b08918032b07963e14b6c94be6ceef, since the
latter just added more code and did not really make the code easier to
read.
It also changes the --or1k-elf-load argument to --elf-load,
to match the rest of the systems.
And finally it fixes a bug in the timeout argument.