wb_intercon: use verilog_utils
Takes the recently added verilog_utils into use,
which make the implementation slightly cleaner.
By this, a bug in the $clog2 implementation in ISE < 14.1
is resolved.
And a workaround for a bug in verilator is removed
(more recent version of verilator than where the bug was
present, is needed for other reasons anyway).